Using "long long" instead of "off_t" in interfaces to guarantee 8-byte data type. off_t, being a typedef, has a different size depending on compiler flags on Solaris.
git-svn-id: https://svn.code.sf.net/p/stemkit/code@124 faf13d3b-af0d-0410-97ba-fcc57a00092e

+++ include/sk/io/AbstractInputStream.h
@@ -29,7 +29,7 @@ namespace sk {
         std::vector<char> read(int number);
         std::vector<char>& read(std::vector<char>& buffer, int number);
         int skip(int number);
-        off_t available() const;
+        long long available() const;
         bool markSupported() const;
         void mark(int readlimit);
         void reset();
